requirements.txt
google_api_python_client==2.188.0
protobuf==6.33.4
Requests==2.32.5
